requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bidict==0.23.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
databases==0.9.0
distro==1.9.0
fastapi==0.114.1
filelock==3.16.1
fsspec==2024.9.0
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.8
jiter==0.5.0
openai==1.44.1
packaging==24.1
passlib==1.7.4
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.4
PyYAML==6.0.2
requests==2.32.3
simple-websocket==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
wsproto==1.2.0
